Ship handles own collision now
The ship called a function that detected whether it or its shot collided with a UFO.  The code for whether it itself collided wasa function owned by the UFO, this is now called by the ship itself.  Perhaps this should be owned by the superclass or handled compeltely separately from each individual object?