Score handling changes
Instead of clumsily detecting if we award the player with score by checking that we are colliding with NOT the player, we now choose whether a particular collision causes an award of score in teh call to detect collisions.  THis means we can now make asteroids, shots and fragments not award any score.