Death handling revamped
Instead of letting all movingwrap children reset, we want to just remove them at death, with the exception of the player.  Keeping with the previous expectation of a set number of enemies on screen we spawn ufos and asteroids up to the expected level.  Missiles are exceptional in that they set an enemy as a target, we need to clear the dead ufo from all missile target variables or if a missile is targetting ufo #1 and #1 dies, the "new" ufo #1 will take its place as the target, which is not what we want.