Elastic collision implemented
Instead of the ship colliding with a ufo or asteroid kiling the ship, there is now an elastic collision with (for now) 10 damage done to the ship and the involved entity.  This function is in movingWrapPositionClass so this can be used just as easily to allow for asteroids colliding with other asteroids, UFOs etc. etc.