Combat text spread out
When we use the beam and/or shockwave weapons, since the damage applies every tick (and ticks happen very fast) the text was creating a blurred mess.  So this function is there to spread out the combat text somewhat.  It doesn't rule out overlap happening and is probbaly horribly inefficient but it works for now...