Shield powerup partial implementation
shield powerup changes colision radius of finder and gives it shieldhp equal to its predefined max shield hp.  Just needs the shield hp to be depleted before regular hp is touched.

Also bug fix: previously we were settign powerup duration for damage and speed to be 300, now we are adding 300 it is neccesary to lock it to a positive number even when decrementing it.