Levels and Waves
Instead of respawning an asteroid or ufo when one dies, we wait until the array for ufos is empty then we spawn new ufos and asteroids based on a configuration for each "wave".  Since asteroids can persist between waves (ufos don't because to trigger a new wave we kill them all...) it means that the player will have to worry about asteroid management more.